admins, bookings, models, added entity type to brands and remove city from entities collection

+++ Modules/AuthModule/Services/AdminService.php
@@ -13,10 +13,14 @@ use Modules\AuthModule\Repositories\Admin\AdminRepository;
 use Modules\AuthModule\Transformers\ActivityResource;
 use Modules\AuthModule\Transformers\AdminResource;
 use Modules\CommonModule\Services\ActivityService;
+use Modules\CommonModule\Traits\ImageHelperTrait;
 use Modules\CommonModule\Transformers\PaginateResource;
+use MongoDB\BSON\ObjectId;
 
 class AdminService
 {
+    use ImageHelperTrait;
+
     private $adminRepository;
 
     public function __construct(AdminRepository $adminRepository)
@@ -26,32 +30,69 @@ class AdminService
 
     public function findAll(Request $request)
     {
-        $admins = $this->adminRepository->list($request, $request->get('limit', 20));
+        $admins = $this->adminRepository->list($request, $request->get('type', 'admin'), $request->get('limit', 20));
 
         return new PaginateResource(AdminResource::collection($admins));
     }
 
-    public function create($request)
+    public function find($adminId)
     {
-        $admin = $this->adminRepository->create([
-            'name'      => $request->name,
-            'email'     => $request->email,
-            'role_id'   => $request->role_id,
-            'password'  => Hash::make($request->password)
-        ]);
+        $admin = $this->adminRepository->find($adminId);
 
         return new AdminResource($admin);
     }
 
-    public function update($adminId, $request)
+    public function create($request)
     {
-        $admin = $this->adminRepository->update($adminId, [
+        $data = [
             'name'      => $request->name,
             'email'     => $request->email,
-            'role_id'   => $request->role_id
-        ]);
+            'role_id'   => new ObjectId($request->role_id),
+            'password'  => Hash::make($request->password),
+            'type'      => $request->type
+        ];
 
-        return new AdminResource($admin);
+        if ($request->type == 'vendor' && $request->vendor_id) {
+            $data['vendor_id'] = new ObjectId($request->vendor_id);
+        }
+
+        if ($request->hasFile('image')) {
+            $data['image'] = $this->uploadImage('avatars', $request->image);
+        }
+
+        $admin = $this->adminRepository->create($data);
+
+        return [
+            'id'    => $admin->id
+        ];
+    }
+
+    public function update($adminId, $request)
+    {
+        $data = [
+            'name'      => $request->name,
+            'email'     => $request->email,
+            'role_id'   => new ObjectId($request->role_id),
+            'type'      => $request->type
+        ];
+
+        if ($request->password) {
+            $data['password'] = Hash::make($request->password);
+        }
+
+        if ($request->type == 'vendor' && $request->vendor_id) {
+            $data['vendor_id'] = new ObjectId($request->vendor_id);
+        }
+
+        if ($request->hasFile('image')) {
+            $data['image'] = $this->uploadImage('avatars', $request->image);
+        }
+
+        $this->adminRepository->update($adminId, $data);
+
+        return [
+            'id'    => $adminId
+        ];
     }
 
     public function updatePassword($adminId, $request)
@@ -60,12 +101,28 @@ class AdminService
             'password'  => Hash::make($request->password)
         ]);
 
-        return new AdminResource($admin);
+        return [
+            'id'        => $adminId
+        ];
     }
 
     public function destroy($adminId)
     {
+        if (auth('admin_api')->id() == $adminId) {
+            return [
+                'message'    => __('action not allowed'),
+                'statusCode' => 400,
+                'errors'     => []
+            ];
+        }
+
         $this->adminRepository->destroy($adminId);
+
+        return [
+            'message'    => __('Data has been deleted successfully'),
+            'statusCode' => 200,
+            'errors'     => []
+        ];
     }
 
     public function getActivities($adminId, $request)
